atoi_simd

Function parse_any

pub fn parse_any<T: Parse>(s: &[u8]) -> Result<(T, usize), AtoiSimdError<'_>>

Parses a slice of digits until it reaches invalid character, and checks the first ‘-’ char for signed integers. Returns the parsed value and the parsed size of the slice.